Changed dgit-locations to timestamp the logfile with the run and other logfile modifications
I just realized that because dgit-locations is the source for the dgitlog variable,
anything that should be inserted before any data is written to the log can be
put in it and will be written to the log. It now also writes the entire bash
callstack (such as it is) to the logfile so that the currently running program
can be easily determined.